made needed changes to fix multi-value case
