+++ src/VecSim/algorithms/hnsw/hnsw_tiered_tests_friends.h
@@ -8,3 +8,4 @@ INDEX_TEST_FRIEND_CLASS(HNSWTieredIndexTest_insertJobAsyncMulti_Test)
 INDEX_TEST_FRIEND_CLASS(HNSWTieredIndexTest_KNNSearch_Test)
 INDEX_TEST_FRIEND_CLASS(HNSWTieredIndexTest_parallelSearch_Test)
 INDEX_TEST_FRIEND_CLASS(HNSWTieredIndexTest_parallelInsertSearch_Test)
+INDEX_TEST_FRIEND_CLASS(HNSWTieredIndexTest_MergeMulti_Test)
