+++ src/VecSim/utils/vec_utils.h
@@ -56,12 +56,13 @@ public:
     static const char *BATCH_SIZE_STRING;
 };
 
+int cmpVecSimQueryResultById(const VecSimQueryResult *res1, const VecSimQueryResult *res2);
+
+int cmpVecSimQueryResultByScore(const VecSimQueryResult *res1, const VecSimQueryResult *res2);
+
 void sort_results_by_id(VecSimQueryResult_List results);
 
 void sort_results_by_score(VecSimQueryResult_List results);
-
-VecSimQueryResult_List merge_results(VecSimQueryResult_List first, VecSimQueryResult_List second,
-                                     size_t limit);
 
 VecSimResolveCode validate_positive_integer_param(VecSimRawParam rawParam, long long *val);
